requirements.txt
torch==1.11.0
mpi4py==3.1.4
scikit-learn==1.0.2
numpy==1.17.3
setproctitle==1.3.3
networkx==2.6.3
tqdm==4.66.1
grpcio==1.59.3
py-diffie-hellman==1.0.1
pycryptodome==3.19.0
certifi==2020.6.20
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
urllib3==1.25.9
flask==1.1.2
gunicorn==20.0.4
paho-mqtt==1.5.0
Celery==4.4.7
dill==0.3.3
torchvision==0.12.0
protobuf==3.17.3
wandb==0.8.36
gevent==20.9.0
greenlet==0.4.17
